+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 2.1.2
+
+- Added: fetch payment ID from PaymentMapping if not provided in order transaction_number
+- Fixed: mark PaymentMapping as captured if charge is captured
+
 ## Version 2.1.1
 
 - Fixed transfers not being created for subscription fees and extra invoices